Refactored project structure, organized routes, controllers
