+++ server/routes.ts
@@ -1,0 +1,22 @@
+import express, {Request, Response} from 'express';
+import {
+  createUser,
+  getAllUsers,
+  deleteUser,
+  updateUser,
+} from './controllers/userControllers';//import controller functions
+
+//create an express router
+const router = express.Router();
+
+//defind routes
+router.get('/', (_req:Request, res:Response) => {
+    res.send('CRUD API');
+});
+router.post('/users', createUser);
+router.get('/users', getAllUsers);
+router.delete('/users/:fname', deleteUser);
+router.put('/users', updateUser);
+
+//export router to be used in server.ts
+export default router;
