+++ users.json
@@ -34,5 +34,9 @@
   {
     "fname": "tungtungtung",
     "lname": "sahur"
+  },
+  {
+    "fname": "Muna",
+    "lname": "achraya"
   }
 ]
